refactor: test case placement and only required test cases
